PHA: move initialization to start of the class
It has bugged me for a while that __init__ is not at the start
of DataPHA, so make it so. It **might** also help with typing
rules: I have seen some complaints that the typing engines can
not work out what some of the fields are meant to be for the
DataPHA class, and these are for properties - that is, some of
the many DatPHA fields where we store the data in a `_xxx`
field and make `xxx` the property getter/setter. I am hopeful
that having the init call first, where some of these internal
fields are set, makes things easier for the typing engines.