DATA: improve docs/typing of id/bkg_id/response_id
Add types for the "id" elements of the API, and some of the
structures it is used in (e.g. a dict of DataPHA or responses).
I have tried to restrict the types to just the IdType values
but a few other cleanups have ended up comnig in. It would be
good to add types for many of the routines, but this is hard
to do (to write, to check, to review). Since python typing
has been written to be "additive" we still add value to the
code base by doing this.

Improve the documentation to better match the "int, str, or None"
behaviour (when relevant).

Note that there is the question of whether the responses have to
be DataARF/RMF - as I have typed them - or can also take ARF/RMF1D
objects. Technically then can (and likely do) take both, since the
instrument versions (ARF/RMF1D) just wrap DataARF/RMF and provide
the same API. However, typing this correctly is awkward, due to
import loops, so leave the types just as DataARF/RMF. One future
change could be to add a protocol to represent the functionality
we need for DataARF/RMF, but that's for later.