UI: improve id documentation/typing
Update sherpa.ui.utils to better type and document the id types
(for this module there is no concern with bkg_id or response_id
arguments, so this is just `id`, `otherid`, and `otherids`).

There is a small amount of code clean up.

The typing attempts to just be related to IdType but there are
some changes made for routines I am changing - e.g. saying they
return None, or a callable that takes an IdType.