UI/ASTRO: improve docs/typing of id/bkg_id/response
Add types for the id elements, then some related types, and others
that are "obvious" as we go along. This is not an attempt to
add full typing to routines as it's surprisingly hard for these
modules, so just do the obvious ones (and do not guarantee that
this catches all the obvious ones).

There is a small amount of code changes, but these are to improve
the typing checks, and do not change the functionality of the code.