DATA: PHA improvements when data is empty/no filter/all filtered
We have a number of very-strange corner-cases with data, particularly
DataPHA, when all the data is filtered out. For example

- pha.mask = False and pha.mask = [False] * nchans
  can lead to different output

- sometimes we error out, sometimes we return None, and
  sometimes we seem to forget there was a filter

- sometimes we return "" and sometimes "No selected channels"

- DataPHA.get_indep defaults to filter=True but
  Data1D.get_indep defaults to filter=False

So this attempts to normalize the code, focussing on DataPHA and the
filter expression, changing things so that it always returns the empty
string (to match other data types) rather than sometimes returning "No
noticed channels".

Partially address #1220.

Note:

  Addressing this too be more consistent has the potential to break
  down-stream code which relied on the previous behaviour.