UI: allow plot/contour layout to be set by the user
Users can either use the new rows and cols arguments to
the plot() and contour() routines to set the value for this
plot, or change the preferences using the return value of
get_split_plot().

The algorithm used to select the number of columns and rows
has now changed (it doesn't force two rows).