TEST: check new plot/contour behaviour
This only checks plot since it is assumed that it will work for
contour as the code is the same in both cases.

Unfortunately, to really test the rows/cols changes, we need to
be able to "test" the resulting plot, which is easiest with
matplotlib (since we don't have an easy way to compare the plot
output, but we can ask programmatically the matplotlib settings).