UI: support overplot for plot, overcontour for contour - fix #2128
This address issue #2128 by allowing

    plot("data", "data", 2)
    plot("model", "model", 2, overplot=True)

We allow the overplot/contour case to add less than the original
number of plots, but we error out if the user wants to display
more.

The SplitPlot class is not well documented, or designed, for
adding data to an existing visualuzation, and so we have had to
access internal variables to get this code to work. Ideally we
would improve the SplitPlot API but

- it's not really clear how we want the class to behave
- the UI code is the only real user of this class at the moment

so I would much rather use this approach until we have worked out
how we want SplitPlot to work.

The current SplitPlot/UI code might also do better separating out
the "suggested" and "being used now" sizes.