TEST: add RMF-offset tests - see issue #2127
Add a number of tests, ranging from

- low-level: sherpa.astro.utils.utils.filter_resp
- object-level: DataRMF
- user level: UI tests
  - these include comparing in-memory datasets to
    ones where the data/responses are written out
    and then read back in

that try to run through tests of the same data but with
offsets of 0, 1, 2, and 5.

The offset=0 tests have been adjusted so that they pass;
this normally means adjusting the set of selected channels
to mimic the current I/O behaviour of renumbering PHA
files with TLMIN<channel>=0 from 0, ..., n - 1 to 1, ..., n.

Note that we don't renumber manually-created PHA files, which
is essentially issue #1305.

Some of the filter_resp (and hence the notice call of DataRMF)
tests are definitely regression tests, as I think they should
return something slightly different. However, I am not sure my
understanding is correct, and I want to make sure we test the
current code before changing anything.