UI: add format and delim arguments to get_filter. Fix #1277
Pass these arguments on to the DataXXX class, if set. Note that
the different DataXXX classes have different defaults - e.g.
DataPHA - so instead of coming up with a compromise, we only
change things if the user specifies them. This means that
existing code will not change (good), but you can now say

    get_filter(format="%.2f")

if you want.