Improve pre-commit script for ranges of years
We often use something like "#  Copyright (C) 2010, 2016 - 2023" and the previous version of the script did not deal with that.

Now, we do. While this is not crucial functionality, is annoyed me so much in development, that I decided that it was faster to fix here than to change the copyright years by hand.
In programming this, I made the choice that even a two year span will be listed as a range (i.e. "2023-2024" instead of "2023, 2024").
Note that this script is opt-in in the sense that developers need to copy it into `.git/hooks` for it to run as an automatic pre-commit step.

fix #1457