UI: add documentation for table to show_xsabund
Closer match to the XSPEC "show abund" command as we now
include the "documentation" for the table as well as the table
name.

This required providing access to the C++ FunctionUtility::abundDoc
routine - fortunately it's easy to wrap and we can do most of the
logic in the Python wrapper.