Address review comments
The C++ API doesn't guarantee the output here - that is if there
is leading or trailing spaces - so update the comment to just say
what it's doing. At the moment (circa XSPEC 12.14.0) it looks
like the C++ code returns a string with a leading space, but this
is not a time-critical routine and I would rather be "safe" in
case the code ever starts adding spaces or unwanted characters
to the end of the string, as opposed to using lstrip() rather
than strip().