TESTS: add corner-case handling of ignore_bad
We do not handle ignore_bad / quality arrays well, so try and
make sure we cover some basic features in case I decide to "fix"
things.

There are some related-ish tests to check what happens with
empty datasets.

These tests are generally written as regression tests - that is,
they check the current behaviour in corner-case situations - because

a) we can find out when things change
b) it's not always obvious what the best thing to do is