DATA: add some guard rails to the quality_filter field
It's not obvious at this time what the semantics of this field
are, so start by adding explicit get/set guards so we can check
that the size/data type is correct, at least.

This does have the advantage of cleaning up a few tests I added
a long while ago which basically said "they should fail but
currently do not, or not nicely".