TESTS: clean up some channel=0 tests
A couple of tests compared the same COUNTS column to the
results when using channel=1,... and channel=0,...

Now, the notice filter - as using channel ranges - should be
different (one less) when using channel=0 to get the same
data out of the calls. **Or** we can keep the same filter
range and check the results are different (but as expected).

Prior to the channel=0 work I was using the latter approach,
but I've switched to the former as it makes it more obvious
that we are getting correct data back.

I've also taken advantage of the offset parameter in
create_delta_rmf - not that it's actually relevant here, but
at least signposts what it's meant to be doing.

There is no functional change in these tests.