DATA: add more typing statements
This tries to add typing statements to all of DataPHA but there are
some hold outs - primarily the exposure, areascal, and backscal
settings, as I do not have the time/energy to work on this just now.

This does not reduce, but rather, increases, the number of complaints
from mypy/pyright since we now have more code we can check. The issues
include, but is not limited to:

- the current API can lead to paths where we know that something is
  valid (i.e. set) but the typing claims it's None | ...

  These could be addressed by adding "assert xxx is not None" lines,
  but it is better to try to impove the API, or change how the code
  is written. This is left for later work.

- the current class structure does not make it easy to type, as
  many of the methods extend the parent class by adding new optional
  keywords, but the parent class doesn't support that.

  Then there's those times when we just "completely" change the
  routine - e.g. the arguments or the return type.

  This again is left for later work.

- there's issues over determining whether something is sequence-like
  or has a size or ... that can throw the type checker into a tizzy
  even though "it is just obvious"