DATA: clean up DataPHA.get_x - fix #2146
So, the DataPHA.get_x() routine calls self._from_channel and is
the only user of this routine (other than some tests). As noted
in #2146 this means that, as get_x always sends in self.channel
to _from_channel we can encode that knowledge in _from_channel
itself.

It turns out that _from_channel is actually one of _group_to_channel,
_channel_to_energy, or _channel_to_wavelength, so we can make
them always work on self.channel. This means we can

a) remove error checks for invalid channel values, as we know
   that can't happen here

b) remove all need for mapping between channel value and index
   number, as we never need to filter the output

   which ends up fixing a bug if channel does not start at 1,
   since the code was taking self.channel - 1 and using that
   as an index.

   The test_get_x_channel0 test in
   sherpa/astro/tests/test_astro_data_swift_unit.py
   checks this out, but it was not written before the initial
   channel=0 fixes, so I can't show the original broken behaviour
   in the test.

It's interesting to ask why the existing tests didn't catch this
change. The reason is two-fold

1. it only affects calls to get_x, and this is only used by
   to_plot

2. however, PHA plots actually bypass the to_plot output
   because
   - we need to deal with different analysis settings and we
     do not have a good API for it, so the code is a bit messy
   - perhaps more importantly, DataPHA derives from Data1D but
     for plotting we actually need Data1DInt-like behaviour,
     so we ignore some of the output from to_plot