Implement cache with more than one canonical value
The new idea here is that instead of using one canonical value which can create problems when the log(absorption) iis either 0 and 1 to numerical precision for high or low values of the energy,
I split the energy into different regions that can be used with different canonical values of N_H.

This requires fine tuning at the level of individual models.
It also requires to loosen up some of the numerical tests, because values are not subtely different, but not how e.g. the fitted values od N_H in one test is essentially the same, it's just he fit statistics that differs slightly.

See https://github.com/hamogu/sherpa-opt-benchmark/blob/main/notebooks/optimize_absorption_cache.ipynb