DATA: clean up DataPHA.get_x
This was requested by @hamogu when reviewing #2118 but at that
time I was worried about the various different options (having
just removed one of the arguments to the routines).

In further work I've had to look at this and realised that the
code never set the group argument - i.e. it always returned
the ungrouped data - which made the code even simpler, and hence
a lot easier to integrate directly into get_x.