DATA: quality_filter cleanup
A number of changes intended to improve handling of data
with a quality filter applied: the idea is to make the
quality filter "always applied" when present

- notice() no longer clears the quality filter, but it will
  clear the other filters

- changing the quality array if ignore_bad has been called will
  now use the new quality array, whereas previously an ignore_bad
  call needed to be called. This is most obvious if you do

    group_counts(5)
    ignore_bad()
    group_counts(4)

This does not fix all problems with quality arrays, and may
make more. There is also the option of requiring ignore_bad to
be done before any filtering calls, but calls like group_counts,
which can change the quality array, complicates things.