UI: note when a 1D filter command has failed
When I originally added the "notice/ignore reports the filter via the
log" feature I explicitly hid the error case. I feared that it would
be confusing for the user. Well, this turns out to be annoying when
there is an error, such as in #1936, as it is very easy to miss the
fact "I should have got a message but did not".

So now we

- handle the case when we've gone from "broken" to "okay" (although to
  be honest the user should still be very wary about this case); this
  is primarily just to separate out the behaviour from the next change

- raise an ERROR message if there is a broken filter

I hope this is a reasonable balance between telling the user something
has gone wrong and not over-loading them with information that doesn't
help them.

I also note that this is ONLY at the UI layer; users of the OO code
have to work this out themselves (but that's been the standard
approach).

Really we should not have these problems but

1. ignore_bad has been a hard nut to crack

2. what do we want to happen when applying a 1D filter to
   2D data

so I think it better to at least warn the users.