XSPEC: add the SpectrumNumber wrapper
The idea is that we want the identifier for the spectrum number not to
be mized-up with other identifiers. This is done by wrapping an
integer into the SpectrumNumber dataclass and making the interface use
this instead of an integer.

This leads to an intended use case where the user doesn't have to
think about the actual number, then can just use next_xsxflt()
to get the next value [*].

One fun thing is that XSPEC models which have the spectrum attribute
now require a SpectrumNumber for input, but it return an int (as the
main time we want to use this is when evaluating the model and we do
not need to worry about the SpectrumNumber abstraction at this time).

[*] this does not need the SpectrumNumber API as it would work
    with integer values, but I think it's clearer.