XSPEC: move per-spectrum logic into a separate interface  WIP
Rather than all models having a spectrum attribute, only those
that use the XSPerSpectrum interface do. This makes it less
confusing for the user, as they only have the .spectrum field
when they need to worry about it.

Whether this is a good idea given our current class structure
is another question.

There is the little fact that it also doesn't work ...