UTILS: replace C++ grouping code with Python
This attempts to replicate the behaviour of the C++ code, in
particular the error handling, but there may be some cases
where there's a difference. The tests pass though.

This code is significantly slower, but it's not obvious to me
that this is a fast code path, since the grouping tends not
to get done during fitting or error analysis (the assumption
is that the code that does the grouping is called before the
data is sent to the fit/optimizer/... code). However, I have
not conclusively proved this.

Here are timings taken for a 8192 channel dataset for one
particular grouping scheme. I would expect different schemes
would cause different run times but I doubt there'd be a
significant difference, apart from pathological cases. I expect
the Python code could be sped up but I think it's quite
readable as is, and only think it's worth changing if we can
show it's worth the effort.

The C++    version comes out as ~ 30 microseconds
    Python                      ~  5 milliseconds

In [13]: %timeit utils._utils.do_group(counts, grouping, "_middle")
26.1 μs ± 661 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [14]: %timeit utils._utils.do_group(counts, grouping, "_middle")
25.8 μs ± 488 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [15]: %timeit utils._utils.do_group(counts, grouping, "_middle")
25.4 μs ± 289 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [16]: %timeit utils.do_group(counts, grouping, "_middle")
4.82 ms ± 7.21 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [17]: %timeit utils.do_group(counts, grouping, "_middle")
4.82 ms ± 13.1 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [18]: %timeit utils.do_group(counts, grouping, "_middle")
4.82 ms ± 5.04 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)