Allow tests to pass with Python 3.13.0 rc3
Python 3.13 has changed how docstrings are indented (in that
it automatically removes leading indentation).

This affects a few of our tests, so update them to allow for the
old and new behaviour. I felt it simpler to just gate on the
Python version (sys.version_info) rather than try to write tests
that could handle both. The downside to my approach is that the
tests repeat themselves. The upside is that it's much easier and
cleaner to update them once Python 3.13 becomes our minimum
supported version.