Allow tests to pass with Python 3.13
Python 3.13 has changed how docstrings are indented (in that
it automatically removes leading indentation).

This affects a few of our tests, so update them to allow for the
old and new behaviour. I felt it simpler to just gate on the
Python version (sys.version_info) rather than try to write tests
that could handle both. The downside to my approach is that the
tests repeat themselves. The upside is that it's much easier and
cleaner to update them once Python 3.13 becomes our minimum
supported version.

There has also been a change in an error message in Python 3.13, and
we have several tests that check for the message. These tests are
meant to be "informtional", in that we know the code is not working at
the moment and we want to know whnever the behaviour changes, either
because the Sherpa code has been changed or the system has changed.

This is an example of the latter. Fortunately we can just check
that the error message is one of the expected forms thanks to
the re module.