BUILD: add a doc build type
With this

    pip install .[doc]

will install most of the code needed for

    cd html
    make html

The remaining elements that need to be installed are pandoc and
(optionally) graphviz, but these are not Python packages.

The constraints on the packages match those in docs/environment.yml,
except that this adds an explicit minimum Sphinx version matching
the constraints currently used by ReadTheDocs.