BUILD: avoid a conversion message when building the docs
In Sphinx 1.8 the source_suffix type was changed from supporting
a string (or list of strings) to also supporting a dict. In a
recent Sphinx release it started to note that it was changing the
setting, so update the configuration to avoid the warning. We
are using Sphinx >= 5 (from RTD and now the doc option) so it
should not cause a problem.