TESTS: Catch error-message change in Python 3.13
An error message has changed in Python 3.13 and we have several
tests that check for the message. These tests are meant to be
"informtional", in that we know the code is not working at the
moment and we want to know whnever the behaviour changes, either
because the Sherpa code has been changed or the system has changed.

This is an example of the latter. Fortunately we can just check
that the error message is one of the expected forms thanks to
the re module.