Review of the Filter.notice call
There should be no functional change here, just some minor style
changes. There has been some discussion of the Python ~ symbol and
I took this as a time to review this section to see what was being
used; as it is a ndarray we can use the slightly-longer but
more-explicit np.invert call.