DataPHA: add get_indep_transform
There hasn't been an "official" way to get at the energy or
wavelength values of a DataPHA object. The only real way to do
it (other than using internal methods) was to create a data plot
and then access the xlo/xhi/x attributes from it.

Actually, that's not fair, since get_x will return values that
match the units setting, **but** it is set up to only return a
single array, and not the low and high edges.

The only other data class that allows for different "units" for
the independent axis is DataIMG, and that class uses the independent
axis directly - that is, img.get_indep() will return the axis values
set by the img.coord value. This approach will not work for DataPHA
without significant refactoring, hence breaking existing code. It
does at least suggest that access via the `indep` attrbuute makes
sense.

So, this adds a `get_indep_transform` call (I had originally tried
adding a `transform` argument to `get_indep` but the two modes are
different enough it is better to keep them separate) for the `DataPHA`
class.

So get_indep()

  - the code behaves as before, so ignores a single array of
    channels
  - the group argument is ignored (we could do something, but it's
    not obvious it is useful so we ignore this)

and get_indep_transform()

  - the code returns a pair of values rather than a singleton that
    give the low and high edges of each bin, in the current
    units setting (so channel, energy, or wavelength)
  - when group=True, filter=True (default) then these are grouped
    and filtered
  - when group=True, filer=False these are just grouped
  - when filter=True, group=False these are just filtered

The idea is that the rest of the code base will be transitioned
to using this method, and then we can mark the existing internal
methods (primarily _get_ebins) can be marked as deprecated.