DataPHA: update get_specresp to use get_indep. Fix #1582
Improve the documentation of get_specresp to mention why
interpolation is needed (basically, I always forget when I
look at this routine so it is useful for me in determining
what is going on).

Issue #1582 noted that the get_specresp call had strange
behaviour when the analysis setting was changed (in particular
for units=channel), and this has been fixed as part of the
transition to using get_indep.

Another issue in #1582 was why the interpolation replaced negative
values with values of 1 rather than 0. This is probably to avoid
division by zero when guess is used, but let's worry about that
at a later time.