DataPHA: update get_filter to use get_indep. Fix #2184
As well as fixing #2184 - that is, the fiter limits are correct
when no data is excluded and the energy scale is in decreasing
order - this also changes the output of get_filter() when data
is excluded because of a bad filter. I have not actually looked
deeply into why this has happened, but I will note that a number
of tests that were changed are regression tests and had notes
along the lines of "this output is technically incorrect, but
we have not been able to fix this". I will note that this is
only a very-small part of improving our "bad" handling, and it
is possible that this is a "false start" (since lots of other
places need to be worked on and this may change the behaviour
here).