DataPHA: update get_x to use get_indep
In doing so I realised that the code uses hc / emid when using
analysis units, rather than (hc / elo + hc / ehi) / 2.  This could be
changed, as it would simplify the get_x implementation, but it is not
clear it is worth it (the DataPHA class does not make much use of
get_x). The documentation has been updated to note this behaviour.