DataPHA: update notice to use get_indep
The replacement code is a bit messy as it explicitly splits up
the channel versus energy/wavelength handling, just to try
and make the separate options clearer. The code could likely
be improved, but leave that for later.

There is some interesting change in behaviour of get_indep_transform
to _get_ebins when all the data has been masked out that we need to
emulate (and perhaps it could go into get_indep)? The real problem is
that the code does not have a consistent approach to dealing with
"filtered out everything", and so it becomes hard to try and use a
single approach as it is likely to break something.