TEST: address numpy 2 changes to output for docstring tests
The str/repr output in NumPy 2 is much-more verbose and no-longet
matches the behaviour of the associated Python types. This breaks
a number of doctests, which had things like

    >>> (a == b).all()
    True
    >>> x.val
    3.4

which now fails for "textual" reasons even though the values are
the same. The simplest approach is just to print the values, so
we now have

    >>> print((a == b).all())
    True
    >>> print(x.val)
    3.4

It is not ideal, but is a simple fix that works with versions 1
and 2 of NumPy.