DataPHA: move checks from set_analysis to the units field
The DataPHA object has set_analysis and get_analysis and the units
field. It was doing validation in set_analysis, but it makes sense
to do this in the units property (as this is the more fundamental
location).

It does mean that some code that used to work no-longer does -
such as

    pha = DataPHA("ex", [1, 2, 3], [3, 4, 5])
    pha.units = "energy"

but I think this is a good thing, as it reduces the number of
invalid states the object can occupy.

As part of this, the behaviour of the background units setting has
subtly changed too. Previously it would only change if a response
was set, but we now always try to change the value (but ignore any
error case).

The __set__state call has had top be changed to suppor this new
behaviour (it needs to make sure certain fields are set up before
the units are set). However, it is not at all opvious if we need to
do this or we can just remove it and __getstate__.