Avoid using bin_lo/hi for grating data. Fix #1564
The Chandra grating files can contain BIN_LO/HI columns for PHA
(and maybe ARF) which can allow the code to switch from channel units
when displaying data without a response. Note that a response is still
needed to actually fit the data.

While this is useful, it makes the code hard to understand, and
there's always a concern over differences between the BIN_XX values
and the actual response. Since this is a "post OGIP" addition that has
not been standardized or picked up by other missions, remove the use
of these columns for mapping between channel and wavelength/energy
units.

As part of this, this removes some response-specific code which tried
to allow the response and model evaluation to use different grids.
This code has never been well described or well tested, and is also a
"post OGIP" change (see issue #1871) so I am happy to see the code be
removed.

This addresses issue #1564 but it does not remove sending around
these values.