TEST: fix test related to issue #800
It looks like there was a typo when loading in the data files and
so this was just testing that a == a. Change so that we load in
a counts and a rate file. There are also some other clean-ups I
made as I was trying to understand the test. These changes are
essentially style changes.