Remove read/write of bin_lo/hi columns for DataPHA/ARF
This essentially completes #1564. The bin_lo/hi remain, but are
basically set to None, in case downstream code is trying to set
them (if trying to use the fields to access data this will break).

The changes focus on DataPHA since this has accessors for the fields,
making it easy to do things like

- ignore any setting
- raise a FutureWarning

The DataARF case just quietly ignores the values, but you can set
them if you want.