INST: code clean up and typing additions
Add typing rules for many, but not all, of the methods and functions.
There are some holes shown by this - e.g. an argument is allowed
to be None but in some cases it appears to be required - that could
be looked at.

There are some subtle changes in the instrument code that could lead
to changes in model evaluation (and hence statistic calculation and
then fit results) as the code had some logic to pick the ARF grid
over the RMF grid (but not consistently). In an OGIP set up they
should be the same, but there's a chance they are subtly different
(that is, numerical issues). The original code appears to be part of
the "allow the grids to be different" set up discussed in issue #1871
that has never been fully developed / tested.

There are also some code cleans ups

- avoid optional values by moving the filter call before the superclass
  __init__ is called (which is likely to call NoNewAttributes...) and
  so avoid having to create a field that starts as None but immediately
  gets set to a ndarray

- we had a lot of lines "a, b = ..." and I have split them into separate
  lines as it does not really change the runtime, and just makes things
  more obvious. None of these were doing a swap - i.e. a, b = 1/b, 1/a"
  and so no need to add in temporary variables. Some of these have also
  been changed to better handle typing changes (e.g. to make sure that
  a variable is not None before using it in an expression).

- related, rather than "x = a; if cond: x = b" write it as
  "if cond: x = b else: x = a"

- move some common code to helper routines (the names are not ideal
  but they are not intended to be used outside externally).

- some simple pylint suggestions