Remove _queue attribute for cache
In modern Python versions, dict itself preserves the order of entries. That means that we do not need to keep a list of cached items around just to record the insertion order.
The savings in memory are insignificant, but this reduces the number of private attributes that caching adds to a class and thus makes it easier to debug in future development.

Fix typos discovered in code review

Co-authored-by: Doug Burke <dburke.gw@gmail.com>