C++: be explicit about the std namespace for the utils code
Label the symbols we use from std explicitly rather than use
"using namespace std;". There are not many in use and I think it
is better to be explicit here. This does affect _utils.cc as well
as utils.hh.

There are still two uses of "using namespace std" but these are
in "external" code:

    sherpa/include/sherpa/MersenneTwister.h
    sherpa/utils/src/sjohnson/Faddeeva.cc