C++: simplify the group template
The _do_group template sent in different types to support the
const and non-const float arrays, but we only ever use them
with the same type, so drop the difference.

The caller to _do_group ensures that the data and group arrays
have the same length, so assume this is a constraint on the
routine and only send in a single scalar.