C++: internal change to how grouped data is looped over
Change 1:

Change the template name from ConstFloatArayType to FloatArrayType.
This is purely syntactic as these templates only accept one array
type for the template.

Change 2:

Rather than loop from start to end - 1 and adding 1 to the index,
loop from start + 1 to end, so there is no need to change the index.
I do not expect this to generate any meaningful change in runtime,
but it just makes the intent clearer (i.e. we say what range we
want to loop over).

The underlying array type we are looping over is a wrapper around
NumPy data, and I do not have the skill or desire to add iterator
support to the Array class to use more-modern (e.g. C++11 idioms)
for looping over the array contents.

Change 3:

Avoid creating a temporary variable for the min and max cases
since we can just use the return value (as is done for the sum
and sum_sq variants).

Change 4:

In _do_group make a minor change to the per-group loop to avoid
an explicit continue statement as we can het the same by adding
an else block. The start parameter could be moved into this
block but it doesn't seem worth it.