C++: add a value_type field to arrays
This allows templates access to the data type, which we take
advantage of in the grouping code to ensure that the group
function has a "sensible" type. Previously this had been hardcoded
as SherpaFloat, which worked because the code is only ever used
with an array of SherpaFloat's, aka double. The new code makes this
explicit.

In reality this should not make a difference since, as mentioned,
the template isn't really used (i.e. the input and output types
are known).

As part of this the grouping code has been updated to use some
C++11 features, in particular the auto keyword. There is no
attempt to force this - e.g. by passing "-std=c++11" to
setuptools.extension.Extenion via extra_compile_args - since
it is probably set up by the caller (via CXXFLAGS, for instance).