Change how invalid limits are checked for
This has been part of the specfit est_method_arg update used for
the Chandra Source Catalog build - e.g.

    https://github.com/sherpa/sherpa/pull/2186

but it is a seperate change from the est_method_arg.

It is not obvios why this was done, but there may be cases where
the tolerance check for limits lead to problems - e.g.

    np.any(proposed < minvals) or np.any(proposed > maxvals)

is a better check than allowing through a proposed value which is
minvals[idx] - delta or maxvals[idx] + delta where

  delta < np.finfo(float).eps

(and given that these values are double, this could happen).