Simplify the package section of pyproject.toml
It looks like we do not need all the package / package-data
setup, although I am not 100% convinced. Let's see how it shakes
out.

I have spent some time trying to remove the src/* files excluded
from the wheel for the extension modules, and it appears that

    "sherpa.x.y.src" = [""]

is the way to do this (since variants like setting ["src/*"] for
"sherpa.x.y" did not work).