TESTS: add examples of bugs #1736 and #2210
The tests are written assuming that sigma should "scale" the output
values. Another option would be to drop the sigma argument, but that
would be a breaking change (although, as the current code doesn't
handle sigma correctly, it's not obvious what sort of a problem it
would be).

Should normal_sample ignore the sigma setting?

Should the ParameterScaleXXX classes ignore the sigma setting?