sim: ParameterScaleMatrix uses the sigma argument
Unlike ParameterScaleVector, there was no attempt to support
the sigma setting in ParameterScaleMatrix. Adding support is
actually a bit interesting, since the code as written assumes
that the optimiser (from the supplied Fit object) provides
a covariance matrix via the "extra_output" field. At present
this is only supported by the LevMar optimizer, and this setting
ignores the "sigma" setting, so we have to manually apply it
(by assuming we can just scale by the sigma value, which should
be fine, but I say this without a statistician's hat to put on).