TEST: can dataspace1d handle channel[0] not being 1 for DataPHA
This is a test of part of issue #2212, and just checks if we can
use dataspace1d (from the astro module) to create channel values
that do not start at 1. At the moment we can not do this, so the
tests are marked as XFAIL.

There are also some other corner-case tests added, such as
does the background handling work, and what happens with simple
"non-sensical" values. Note that this last case does not always
lead to an error.