UI: astro dataspace1d DataPHA improvements
The start, stop, step, and numbins arguments are now used when
creating a DataPHA object. This means

- you can now create a channel range that does not start at
  1 (see issue #2212)
- if the resulting channel range is not s, s + 1, s + 2, ..., s + n
  with s being an integer then there is an error

Previously you could say

   dataspace1d(23, 27, step=2, dstype=DataPHA)

and this would create a channel range of

   [1, 2, 3]

which is a bit strange (I can understand the logic now, but it's
hard to explain). So we now error out. Since some errors come
from sherpa.utils.dataspace1d we can't have a single error class
(either DataErr or ValueError), so we have either TypeError or
DataErr, depending on what the error is.

We could extend error checks since

    dataspace1d(1, 5.1, dstype=DataPHA)

creates a channel grid of [1, 2, 3, 4, 5, 6], but

    dataspace1d(1.1, 5, dstype=DataPHA)

errors out, but I don't think it's worth it.